The first version of the deck editor.
The label.lua has been edited to support mouse listener attaching.
The darius_card_pool.lua has been edited to include a few modifications to the interfacing between it and the deck editor.
And finally the darius_deck_editor.lua contains the first version of the deck editor, you can activate and try it trough the widget menu on F11, but there really isn't that much to try...